refactor:: 즐겨찾기 삭제 기능 구현
