+++ src/main/java/nextstep/subway/domain/exceptions/CanNotDeleteFavoriteException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.exceptions;
+package nextstep.subway.domain.exceptions;
 
 public class CanNotDeleteFavoriteException extends RuntimeException {
     public CanNotDeleteFavoriteException(String message) {
